📦 chore(manager.py): add DatabaseManager class to handle database operations 📦 chore(utils.py): add initialize_database function and session_getter context manager to handle database initialization and session management
